refactor: updated button -> withButton and exported it to index
